fix: adding console log data too
